final changes

+++ cowin/Models/Session.cs
@@ -10,7 +10,7 @@ namespace cowin.Models
     {
         public string session_id { get; set; }
         public string date { get; set; }
-        public int available_capacity { get; set; }
+        public decimal available_capacity { get; set; }
         public int min_age_limit { get; set; }
         public string vaccine { get; set; }
         public List<string> slots { get; set; }
